refactor: transform localization module to TS
